% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbem_run_check.R
\name{dbem_run_check}
\alias{dbem_run_check}
\title{Double check DBEM runs}
\usage{
dbem_run_check(results_path = NA, dbem_path = NA, settings = "Settings1F")
}
\arguments{
\item{results_path}{path where results are if not on personal scracth folder}

\item{dbem_path}{path where the DBEM inout files exist}

\item{settings}{settings file for DBEM run. Preallocation Settings1F}
}
\description{
This function double checks all runs from the DBEM are completed correctly
within Compute Canada. It cross-references the results files with
the settings and sbem inpout files. Specifically, the number of species ran,
that each species has the correct number of years, and will output that info.
}
\author{
Juliano Palacios Abrantes | j.palacios@oceans.ubc
}
