% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_dbem.R
\name{grid_dbem}
\alias{grid_dbem}
\title{Matches shapefile with DBEM grid}
\usage{
grid_dbem(shapefile, dbem_grid, out = "df")
}
\arguments{
\item{shapefile}{can be a shapefile or a df. if DF make sure longitude and latitude are present as "lon and "lat}

\item{dbem_grid}{the DBEM grid}

\item{out}{options of output df = for data frame sf for shapefile}
}
\value{
A datafrmae with the
}
\description{
Matches shapefile with DBEM grid
}
\author{
Juliano Palacios Abrantes | j.palacios@oceans.ubc
}
