% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dbem.R
\name{read_dbem}
\alias{read_dbem}
\title{Reads DBEM files}
\usage{
read_dbem(
  taxon_key,
  year,
  cmip,
  model = "All",
  rcp,
  data_type,
  root_path,
  my_path = FALSE
)
}
\arguments{
\item{taxon_key}{is the species to load as taxon key number.}

\item{year}{expects a sequence of years to load the data from}

\item{cmip}{can be 5 for CMIP5 or 6 CMPI6. Note that CMPI5 only has GFDL, IPSL and MPI. CMPI6 has additional CNRM and UKESM}

\item{model}{is the ESM models to load; GFDL; IPSL, MPI, CNRM, UKESM. For all models select "All";}

\item{rcp}{expects "26" for RCP 2.6-low emission scenario and "85" for RCP 8.5-high emission scenario}

\item{data_type}{select Catch or Abd data}

\item{root_path}{Expects the root computer path before DROBO. Note: do not include /}

\item{my_path}{If you have a completely different path for the data. Note that you still have to respect "model/rcp/taxonkey/" structure}
}
\value{
A tidy table -tiddy- with the DBEM data selected
}
\description{
This function loads the DBEM files from DROBO. You have to be connected to the UBC network
and have access to the CORU-DROBO in order to use the function. Note that no user/id will be
required but the path wont be found.
}
