% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_distribution.R
\name{read_distribution}
\alias{read_distribution}
\title{Reads distribution of taxons}
\usage{
read_distribution(taxon_key, root_path = NA, output = "plot", coords = NA)
}
\arguments{
\item{taxon_key}{is the species to load as taxon key number.}

\item{root_path}{Expects the root computer path before DROBO. Note: do not include /}

\item{box}{Expects a vector with four values in the following order: low lat, high lat, low long and high long to load geographical specific data.. if FALSE it will load global database}
}
\value{
A data frame or a map showing the distribution of the taxa

#' @export
}
\description{
This function loads the distribution of a taxon from DROBO. You have to be connected to the UBC network
and have access to the CORU-DROBO in order to use the function. You can choose to return the data or just a plot for visualization Note that no user/id will be
required but the path wont be found. Note that this function needs/lodas the CORU coordinate grid
}
