% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_plot.R
\name{summary_clim}
\alias{summary_clim}
\title{Reads ocean variables from ESMs}
\usage{
summary_clim(cmip, model, rcp, variable, years, root_path, box = FALSE,
  plot_type)
}
\arguments{
\item{cmip}{can be 5 for CMIP5 or 6 CMPI6. Note that CMPI5 only has GFDL, IPSL and MPI. CMPI6 has additional CNRM and UKESM}

\item{model}{is the ESM models to load; GFDL; IPSL, MPI, CNRM, UKESM. For all models select "All";}

\item{rcp}{expects "26" for RCP 2.6-low emission scenario and "85" for RCP 8.5-high emission scenario}

\item{variable}{expects a variable to be load}

\item{years}{expects a sequence of years to load the data from}

\item{box}{Expects a vector with four values in the following order: low lat, high lat, low long and high long to load geographical specific data.. if FALSE it will load global database}

\item{plot_type}{determines the type of plot you want. Options include: time_line, a timeline of the environ data (geom_line)}
}
\description{
This function loads reads oceanic climate variables from DROBO. You have to be connected to the UBC network
and have access to the CORU-DROBO in order to use the function. Note that no user/id will be
required but the path wont be found.
}
\author{
Juliano Palacios Abrantes | j.palacios@oceans.ubc
}
