% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_test_run.R
\name{vb_test_run}
\alias{vb_test_run}
\title{Checking if the VB protocol worked properly}
\usage{
vb_test_run(yr, model, ssp, variable)
}
\arguments{
\item{yr}{A set of years to run the function}

\item{model}{The Earth System model that will be running. Part of the path. Needs to match the file name}

\item{ssp}{The SSP to run. Part of the path. Needs to match the file name}

\item{variable}{Expects the name of the variable to be converted. Part of the path. Needs to match the file name}
}
\value{
Returns three plots to visualize differences; a histogram, a historic line and a map of 10 years average
}
\description{
This function compares the data before and after the VB run
to make sure it worked. It only provides a visual comparison
to see if the extrapolation was effective. All generated plots,
need to be different from each other, otherwise something went
wrong with the VB simulation. Expects the processed data to be
saved in the same root that raw data within a folder called
Processed720_Annualaverage_txt. File names must be the same.
}
