% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_totalphy2.R
\name{vb_totalphy2}
\alias{vb_totalphy2}
\title{Estimates totalphi2 for DBEM
#'}
\usage{
vb_totalphy2(yr, model, ssp)
}
\arguments{
\item{model}{The Earth System model that will be running. Part of the path. Needs to match the file name}

\item{ssp}{The SSP to run. Part of the path. Needs to match the file name}

\item{years}{A set of years to run the function}

\item{save_path}{Expects a path to save the data. Needs overwrite = F. Do not include / at the end of the path}
}
\value{
It does not return anything. This function will overwrite the original file unless the parameter overwrite = F
}
\description{
Estimates totalphi2 for DBEM
#'
}
\author{
Juliano Palacios Abrantes | j.palacios@oceans.ubc
}
