% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clim.R
\name{read_clim}
\alias{read_clim}
\title{Reads ocean variables from ESMs}
\usage{
read_clim(
  cmip = 6,
  esm,
  rcp,
  variable,
  years,
  root_path,
  box = FALSE,
  coords = FALSE
)
}
\arguments{
\item{cmip}{can be 5 for CMIP5 or 6 cmip6. Note that CMPI5 only has GFDL, IPSL and MPI. CMPI6 has additional CNRM and UKESM}

\item{esm}{is the ESM esms to load; GFDL; IPSL, MPI, CNRM, UKESM. For all esms select "All";}

\item{rcp}{expects "26" for RCP 2.6-low emission scenario and "85" for RCP 8.5-high emission scenario}

\item{variable}{expects a variable to be load. Options include: "AdvectionU","AdvectionV", "bot_temp", "htotal_btm", "htotal_surf", "IceExt", "O2_btm","O2_surf","Salinity_btm","Salinity_surf","SST","totalphy2". Note: if you type "options" the function will print all available variables}

\item{years}{expects a sequence of years to load the data from}

\item{box}{Expects a vector with four values in the following order: low lat, high lat, low long and high long to load geographical specific data.. if FALSE it will load global database}

\item{coords}{This is the DBEM/CORU coordinate grid with index, lon and lat}
}
\value{
A tibble with the environmental data requested for the world or a specific box
}
\description{
This function reads oceanic climate variables from DROBO. You have to be connected to the UBC network
and have access to the CORU-DROBO in order to use the function. Note that no user/id will be
required but the path wont be found.
}
\author{
Juliano Palacios Abrantes | j.palacios@oceans.ubc
}
