% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_test_vars.R
\name{vb_test_vars}
\alias{vb_test_vars}
\title{Check VB variables}
\usage{
vb_test_vars(model)
}
\arguments{
\item{model}{The Earth System model that will be running. Part of the path. Needs to match the file name}
}
\value{
It returns a message informing the completion of the VB runs
}
\description{
This part double checks that the VB runs were successful.
Specifically, it makes sure you VB all variables, ssps and years.
It will print out missing data.
}
\author{
Juliano Palacios Abrantes | j.palacios@oceans.ubc
}
